requirements.txt
pymax
aiogram
python-telegram-bot
